Add DelegateTo impl type
