+++ crates/cgp-component/src/lib.rs
@@ -6,6 +6,7 @@
 */
 
 pub mod traits;
+pub mod types;
 
 pub use cgp_component_macro::{define_components, delegate_components, derive_component};
 pub use traits::{DelegateComponent, HasComponents};
