+++ crates/cgp-component/src/types/delegate.rs
@@ -1,0 +1,3 @@
+use core::marker::PhantomData;
+
+pub struct DelegateTo<Components>(pub PhantomData<Components>);
