+++ crates/cgp-error/src/lib.rs
@@ -1,7 +1,5 @@
 mod can_raise_error;
-mod delegate_error;
 mod has_error_type;
 
 pub use can_raise_error::*;
-pub use delegate_error::*;
 pub use has_error_type::*;
