+++ crates/cgp-type/src/lib.rs
@@ -1,3 +1,4 @@
 #![no_std]
 
+pub mod impls;
 pub mod traits;
